+++ django_domain_events/codecs/__init__.py
@@ -1,5 +1,6 @@
 from django_domain_events.codecs.dataclass_codec import DataclassCodec
 from django_domain_events.codecs.payload_codec import PayloadCodec
+from django_domain_events.codecs.payload_encoder import PayloadEncoder
 from django_domain_events.codecs.unsupported_payload_type import UnsupportedPayloadType
 
-__all__ = ["DataclassCodec", "PayloadCodec", "UnsupportedPayloadType"]
+__all__ = ["DataclassCodec", "PayloadCodec", "PayloadEncoder", "UnsupportedPayloadType"]